Update README.md
Trying to fix error.  The screenshot display on my computer, but they do not display when you open the GitHub repo and view the README.MD. The screenshots in the README.MD file are not showing. The images are in the folder, they exist.  I've tried changing folder name (Asset vs asset) and the image names. 

When I click on the image links it says:
 
"404 - page not found
The main branch of Mod3-Employee_Payroll_Tracker does not contain the path assets/images/consoleAndTrackerScreenshot.png."